Change some data sizes from 32-bits to 64-bits, to allow for larger face textures.
There are several places in the code where 32-bit sizes are used that
need to be 64-bit in order to handle large textures.

+++ src/ptex/PtexUtils.h
@@ -177,10 +177,10 @@ void deinterleave(const void* src, int sstride, int ures, int vres,
                   void* dst, int dstride, DataType dt, int nchannels);
 
 PTEXAPI
-void encodeDifference(void* data, int size, DataType dt);
+void encodeDifference(void* data, size_t size, DataType dt);
 
 PTEXAPI
-void decodeDifference(void* data, int size, DataType dt);
+void decodeDifference(void* data, size_t size, DataType dt);
 
 typedef void ReduceFn(const void* src, int sstride, int ures, int vres,
                       void* dst, int dstride, DataType dt, int nchannels);
